+++ src/three/renderer/tiles/raycastTraverse.js
@@ -13,7 +13,7 @@ function distanceSort( a, b ) {
 
 function intersectTileScene( tile, raycaster, renderer, intersects ) {
 
-	const { scene } = tile.cached;
+	const { scene } = tile.engineData;
 	const didRaycast = renderer.invokeOnePlugin( plugin => plugin.raycastTile && plugin.raycastTile( tile, scene, raycaster, intersects ) );
 	if ( ! didRaycast ) {
 
@@ -66,7 +66,7 @@ export function raycastTraverseFirstHit( renderer, tile, raycaster, localRay = n
 		}
 
 		// track the tile and hit distance for sorting
-		const boundingVolume = child.cached.boundingVolume;
+		const boundingVolume = child.engineData.boundingVolume;
 		if ( boundingVolume.intersectRay( localRay, _vec ) !== null ) {
 
 			_vec.applyMatrix4( group.matrixWorld );
@@ -140,7 +140,7 @@ export function raycastTraverse( renderer, tile, raycaster, intersects, localRay
 	}
 
 	const { group, activeTiles } = renderer;
-	const { boundingVolume } = tile.cached;
+	const { boundingVolume } = tile.engineData;
 
 	// get the ray in the local group frame
 	if ( localRay === null ) {